+++ Sources/TuistLoader/Utils/ManifestFilesLocator.swift
@@ -7,12 +7,20 @@ public protocol ManifestFilesLocating: AnyObject {
     /// It locates the manifest files that are have a connection with the
     /// definitions in the current directory.
     /// - Parameter at: Directory for which the manifest files will be obtained.
-    func locate(at: AbsolutePath) -> [(Manifest, AbsolutePath)]
+    func locateProjectManifests(at: AbsolutePath) -> [(Manifest, AbsolutePath)]
 
     /// It locates all manifest files under the root project folder
     /// - Parameter at: Directory for which the **project** manifest files will
     ///                 be obtained
-    func locateAll(at: AbsolutePath) -> [(Manifest, AbsolutePath)]
+    func locateAllProjectManifests(at: AbsolutePath) -> [(Manifest, AbsolutePath)]
+
+    /// It traverses up the directory hierarchy until it finds a `Config.swift` file.
+    /// - Parameter at: Path from where to do the lookup.
+    func locateConfig(at: AbsolutePath) -> AbsolutePath?
+
+    /// It traverses up the directory hierarchy until it finds a `Setup.swift` file.
+    /// - Parameter at: Path from where to do the lookup.
+    func locateSetup(at: AbsolutePath) -> AbsolutePath?
 }
 
 public final class ManifestFilesLocator: ManifestFilesLocating {
@@ -23,7 +31,7 @@ public final class ManifestFilesLocator: ManifestFilesLocating {
         self.rootDirectoryLocator = rootDirectoryLocator
     }
 
-    public func locate(at: AbsolutePath) -> [(Manifest, AbsolutePath)] {
+    public func locateProjectManifests(at: AbsolutePath) -> [(Manifest, AbsolutePath)] {
         Manifest.allCases.compactMap { manifest in
             let path = at.appending(component: manifest.fileName(at))
             if FileHandler.shared.exists(path) { return (manifest, path) }
@@ -31,10 +39,34 @@ public final class ManifestFilesLocator: ManifestFilesLocating {
         }
     }
 
-    public func locateAll(at: AbsolutePath) -> [(Manifest, AbsolutePath)] {
-        guard let rootPath = rootDirectoryLocator.locate(from: at) else { return locate(at: at) }
+    public func locateAllProjectManifests(at: AbsolutePath) -> [(Manifest, AbsolutePath)] {
+        guard let rootPath = rootDirectoryLocator.locate(from: at) else { return locateProjectManifests(at: at) }
         let projectsPaths = FileHandler.shared.glob(rootPath, glob: "**/\(Manifest.project.fileName(at))").map { (Manifest.project, $0) }
         let workspacesPaths = FileHandler.shared.glob(rootPath, glob: "**/\(Manifest.workspace.fileName(at))").map { (Manifest.workspace, $0) }
         return projectsPaths + workspacesPaths
+    }
+
+    public func locateConfig(at: AbsolutePath) -> AbsolutePath? {
+        let subPath = RelativePath("\(Constants.tuistDirectoryName)/\(Manifest.config.fileName(at))")
+        return traverseAndLocate(at: at, appending: subPath)
+    }
+
+    public func locateSetup(at: AbsolutePath) -> AbsolutePath? {
+        let subPath = RelativePath(Manifest.setup.fileName(at))
+        return traverseAndLocate(at: at, appending: subPath)
+    }
+
+    // MARK: - Helpers
+
+    private func traverseAndLocate(at: AbsolutePath, appending subpath: RelativePath) -> AbsolutePath? {
+        let manifestPath = at.appending(subpath)
+
+        if FileHandler.shared.exists(manifestPath) {
+            return manifestPath
+        } else if at != .root {
+            return traverseAndLocate(at: at.parentDirectory, appending: subpath)
+        } else {
+            return nil
+        }
     }
 }